Do not check for available commands in deploy.sh
Since using the display manager on NixOS does not rely on the existence
of startx, it is rather pointless to place the xorg related symlinks
only if startx is present on the system. Therefore, just dump create the
symlinks no matter if xorg or some exotic shell is installed or not.

Same warning as in aprevious commit applies: The script will overwrite
any existing files when placing symlinks. Make sure to backup them
yourself until a backup mechanism has been implemented.